Turned the ShiDB-2.0 module into a maven project to make dependencies easier (this was way harder than expected)
# Conflicts:
#	ShiDB-2.0/test/file/FileMgrTest.java

+++ ShiDB-2.0/src/main/java/log/LogMgr.java
@@ -45,7 +45,7 @@ public class LogMgr {
         lastSavedLSN = latestLSN;
     }
 
-    public void flush(int lsn) {
+    public void flush(long lsn) {
         if (lsn >= lastSavedLSN)
             flush();
     }
